+++ infrastructure/queue/NotifySubscribersJob.php
@@ -23,6 +23,9 @@ final class NotifySubscribersJob extends BaseObject implements Job, RetryableJob
     {
         $logger = new YiiPsrLogger('sms');
         $jobQueue = Yii::$app->get('queue');
+        // TODO: здесь юзается Service Locator, зависимости скрыты.
+        // По-хорошему надо разделять Job (DTO) и JobHandler (DI).
+        // Но yii2-queue заставляет писать логику в джобе, так что живем с этим.
         $queryService = Yii::$container->get(SubscriptionQueryService::class);
 
         $message = Yii::t('app', 'New book released: {title}', ['title' => $this->title]);